+++ src/compare-page-pull-request/index.js
@@ -1,0 +1,1 @@
+export { default } from './compare-page-pull-request';